modified:   lib/drawer_options/lpar_config.dart 	modified:   lib/drawer_options/multiplayer.dart 	modified:   lib/drawer_options/select_gecko.dart 	modified:   lib/main.dart 	modified:   lib/patch_window.dart 	modified:   lib/settings.dart 	modified:   lib/utils/log_utils.dart

+++ lib/drawer_options/lpar_config.dart
@@ -35,7 +35,9 @@ class _LparConfigState extends State<LparConfig> {
   }
 
   void resetLparFile(String packPath) {
-    File(path.join(packPath, 'lpar.txt')).deleteSync();
+    if (File(path.join(packPath, 'lpar.txt')).existsSync()) {
+      File(path.join(packPath, 'lpar.txt')).deleteSync();
+    }
     createLparFile(packPath);
   }
